Combine ethereum/web3 and ethereum/wallet states
