+++ user-registry-ui/ethereum/selectors.js
@@ -1,0 +1,4 @@
+export const getWeb3Id = state => state.ethereum.web3.id
+export const getEthereumAccount = state => state.ethereum.account
+export const getError = state => state.ethereum.error
+export const isConnected = state => state.ethereum.connected
