--- user-registry-ui/ethereum/wallet/index.js
@@ -1,3 +1,0 @@
-export * from './actions'
-export * from './selectors'
-export * from './reducer'
