--- user-registry-ui/ethereum/wallet/selectors.js
@@ -1,3 +1,0 @@
-export const getEthereumAccount = state => state.ethereum.wallet.account
-export const getError = state => state.ethereum.wallet.error
-export const isConnected = state => state.ethereum.wallet.connected
