--- user-registry-ui/ethereum/web3/actions.js
@@ -1,3 +1,0 @@
-import { createAction } from '@reduxjs/toolkit'
-
-export const newWeb3Id = createAction('web3/id/new')
