--- user-registry-ui/ethereum/web3/index.js
@@ -1,4 +1,0 @@
-export * from './selectors'
-export * from './reducer'
-export * from './actions'
-export * from './getWeb3'
