--- user-registry-ui/ethereum/web3/reducer.js
@@ -1,12 +1,0 @@
-import { createReducer, combineReducers } from '@reduxjs/toolkit'
-import { newWeb3Id } from './actions'
-
-const uniqueId = Math.random
-
-const idReducer = createReducer(uniqueId(), {
-  [newWeb3Id]: () => uniqueId()
-})
-
-export const web3Reducer = combineReducers({
-  id: idReducer
-})
