--- user-registry-ui/ethereum/web3/selectors.js
@@ -1,1 +1,0 @@
-export const getWeb3Id = state => state.ethereum.web3.id
